move everything to pkg/web
