+++ go.mod
@@ -265,6 +265,4 @@ replace github.com/crewjam/saml => github.com/grafana/saml v0.0.0-20211007135653
 
 replace github.com/apache/thrift => github.com/apache/thrift v0.14.1
 
-replace gopkg.in/macaron.v1 => ./pkg/macaron
-
 replace github.com/hashicorp/consul => github.com/hashicorp/consul v1.10.2
