--- pkg/macaron/go.mod
@@ -1,11 +1,0 @@
-module gopkg.in/macaron.v1
-
-go 1.17
-
-require github.com/stretchr/testify v1.7.0
-
-require (
-	github.com/davecgh/go-spew v1.1.0 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
-)
