+++ pkg/web/binding_test.go
@@ -1,4 +1,4 @@
-package macaron
+package web
 
 import (
 	"errors"
